feat(filters): Export getApplicableFilters
